Update license to SPDX format, remove license from classiefier
We were having issues with installing the library due to licensing issues. Seems a pip update breaks support for arbitrary licensing strings.

I've updated the license field to match SPDX identifier for the license in the LICENSE file[1].

The licence in the classifiers was causing a `SetuptoolsDeprecationWarning`. Removing it solves that error.

[1] https://spdx.org/licenses/BSD-2-Clause.html

Signed-off-by: Paul Bersee <paul.bersee@prodrive-technologies.com>